파일 분리화
