+++ src/test/kotlin/lottery/LotteryGameTest.kt
@@ -1,0 +1,6 @@
+package lottery
+
+import io.kotest.core.spec.style.StringSpec
+
+class LotteryGameTest : StringSpec({
+})
